requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# AI Integration
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Documentation
pydantic[email]==2.5.3
